before megring the animal_livewire
