+++ resources/views/dashboard.blade.php
@@ -40,7 +40,7 @@
     <div class="flex flex-col items-center">
         <livewire:Animal :animals="$animals" />
         <livewire:Shelter :animals="$animals" />
-        <livewire:Vets/>
+        <livewire:VetCard/>
     </div>
     <script>
         function headerComponent() {
